Change to 3 Layer Folder Structure, Large File Refactor (Config file, etc), INTRODUCED BUG: **Login/Main Session Redirect Doesn't Work Now**
