+++ business/js/printStat.js
@@ -1,30 +1,30 @@
 function printStatus(status) {
     this.statusBox = document.getElementById('statusBox');
-    if (status == "signupU") { //Signup Username already taken
+    if (status === "signupU") { //Signup Username already taken
         this.statusBox.innerHTML = "<span class=\"fail\">Username Already Taken</span>";
     }
-    if (status == "signupE") {
+    if (status === "signupE") {
         this.statusBox.innerHTML = "<span class=\"fail\">A user account with that email already exists</span>"
     }
-    if (status == "singupD") { //Database Error!
+    if (status === "signupD") { //Database Error!
         this.statusBox.innerHTML = "<span class=\"fail\">Database Error</span>";
     }
-    if (status == "loginF") { //Login Fail
+    if (status === "loginF") { //Login Fail
         this.statusBox.innerHTML = "<span class=\"fail\">Username or Password is Invalid</span>";
     }
-    if (status == "notA") { //Not Admin
+    if (status === "notA") { //Not Admin
         this.statusBox.innerHTML = "<span class=\"fail\">You are not admin</span>";
     }
-    if (status == "changePassE") {
+    if (status === "changePassE") {
         this.statusBox.innerHTML = "<span class=\"fail\">Email does not exist</span>";
     }
-    if (status == "changePassD") {
+    if (status === "changePassD") {
         this.statusBox.innerHTML = "<span class=\"fail\">Database Error</span>";
     }
-    if (status == "changePassS") {
+    if (status === "changePassS") {
         this.statusBox.innerHTML = "<span class=\"success\">Successfully changed password</span>";
     }
-    if (status == "login") {
+    if (status === "login") {
         this.statusBox.innerHTML = "<span class=\"fail\">You must login to access this content</span>"
     }
 }
