+++ business/login/logoutScript.php
@@ -1,0 +1,7 @@
+<?PHP
+
+header_remove();
+session_start();
+session_destroy();
+require_once('../../config/config.php');
+header("Location: ".PATH_INDEX);
