+++ config/config.php
@@ -1,0 +1,25 @@
+<?PHP
+
+// Server Configuration Environment Variables
+const DB_SERVER = 'localhost';
+const DB_USER = 'root';
+const DB_PASS = '';
+const DB_NAME = 'school';
+
+// File Paths
+//const PATH_PRINT_STAT = '../../business/js/printStat.js';
+//const PATH_SIGNUP = '../../business/js/signup.js';
+const PATH_LOGOUT_SCRIPT = '../../business/login/logoutScript.php';
+const PATH_PW_UTILS = '../../business/login/passwordUtils.php';
+const PATH_CHANGE_PW_SCRIPT = '../../data/login/changePasswordScript.php';
+const PATH_LOGIN_SCRIPT = '../../data/login/loginScript.php';
+const PATH_REG_ADMIN_SCRIPT = '../../data/login/registerAdminScript.php';
+const PATH_REG_USER_ACCOUNT_SCRIPT = '../../data/login/registerUserAccountScript.php';
+const PATH_CHANGE_PW = '../../presentation/login/changePassword.php';
+const PATH_LOGIN = '../../presentation/login/login.php';
+const PATH_REG_ADMIN = '../../presentation/login/registerAdmin.php';
+const PATH_REG_USER_ACCOUNT = '../../presentation/login/registerUserAccount.php';
+const PATH_MAIN_PAGE = '../../presentation/mainPage/mainPage.php';
+const PATH_STYLES_LOGIN = '../../styles/login.css';
+const PATH_STYLES_STYLE = '../../styles/style.css';
+const PATH_INDEX = '../../index.html';
