+++ index.html
@@ -4,14 +4,14 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="./styles/style.css">
     <title>Temp Home Page</title>
 </head>
 
 <body>
 	<p>Temp Home Page</p>
-    <button onclick="location.href='./login/login.php'">Login Page</button>
-    <button onclick="location.href='./mainPage/mainPage.php'">Main Page</button>
+    <button onclick="location.href='presentation/login/login.php'">Login Page</button>
+    <button onclick="location.href='presentation/mainPage/mainPage.php'">Main Page</button>
 </body>
 
 </html>
