--- login/configure.php
@@ -1,8 +1,0 @@
-<?PHP
-
-define('DB_SERVER', 'localhost');
-define('DB_USER', 'root');
-define('DB_PASS', '');
-define('DB_NAME', 'school');
-
-?>
