--- login/logoutScript.php
@@ -1,6 +1,0 @@
-<?PHP
-	header_remove(); 
-	session_start();
-	session_destroy();
-	header('Location: ../index.html');
-?>
